feat: allow creating Rows from a RecordBatch
This change modifies the `Rows` struct to allow it to be created from an
`arrow_flight::RecordBatch`. This provides a more efficient way to use
the `BulkStreamWriter` when data is already in `RecordBatch` format.

The `Rows` struct was refactored to hold either a builder for
incremental row creation or a pre-existing `RecordBatch` in a single
`RowsData` enum to enforce mutual exclusivity. A new constructor
`Rows::from_record_batch` has been added, and all relevant methods have
been updated to handle the two internal states. New error types have
been added to ensure type safety and prevent invalid operations.

Signed-off-by: Matthias Wende <matthias.wende@frequenz.com>